Wrapper cleanup #138 (#178)
* Remove redundant is_equal_to

* Fix squad_params return type

* config select take int index

* setVariable with public flags

* count(config) return int

* world_to_screen return optional

* added side_get and group_get alternatives

* fix

* squad_params use typedef

* Implement some new wrappers

* fix

* addforce/addtorque

* backpack_space_for

* vectorModel commands

* mousePos/uiContext/missionVersion

* fix

* Bad appveyor

+++ src/client/headers/client/sqf/config.hpp
@@ -42,7 +42,7 @@ namespace intercept {
         std::vector<config> config_properties(const config &config_entry, sqf_string_const_ref condition_ = "true", bool inherit = true);
         sqf_return_string config_source_mod(const config &config_entry_);
         sqf_return_string_list config_source_mod_list(const config &config_entry_);
-        float count(const config &config_entry_);
+        int count(const config &config_entry_);
         //std::vector<game_value> get_array(const config &config_entry_);
         config get_mission_config(sqf_string_const_ref value_);
         float get_number(const config &config_entry_);
@@ -54,7 +54,7 @@ namespace intercept {
         bool is_number(const config &config_entry_);
         bool is_text(const config &config_entry_);
         std::vector<config> config_classes(sqf_string_const_ref condition_, const config &config_);
-        config select(const config &a_config_, float a_number_);
+        config select(const config &a_config_, int a_number_);
         config campaign_config_file();
         config config_file();
         config config_null();
